fix: split dispatch return ABIs
