+++ crates/evm2/src/interpreter/instructions/table/normal.rs
@@ -1,29 +1,27 @@
 use super::InspectMode;
 use crate::{
     EvmConfig, EvmConfigSelector, EvmTypes, SpecId, VersionTables,
-    constants::STACK_LIMIT,
     evm::config::SelectorVersionTables,
-    interpreter::{InterpreterState, Pc, Result, Stack, gas::RemainingGas},
 };
-use core::hint::cold_path;
 
-/// Normal instruction return value.
-#[cfg(not(dispatch_single_return))]
-pub(crate) type InstrFnRet = (Pc, GasStackLen);
+cfg_if::cfg_if! {
+    if #[cfg(dispatch_single_return)] {
+        mod single_return;
+        use single_return as imp;
+    } else if #[cfg(dispatch_packed)] {
+        mod packed;
+        pub(crate) use packed::unpack_ret;
+        use packed as imp;
+    } else {
+        mod unpacked;
+        use unpacked as imp;
+    }
+}
 
-/// Normal instruction return value.
-#[cfg(dispatch_single_return)]
-pub(crate) type InstrFnRet = DispatchResult;
+use imp::dispatch;
 
 /// Normal instruction function pointer.
-pub(super) type RawInstrFn<T> = extern_table!(
-    fn(
-        pc: Pc,
-        stack: Stack<'_>,
-        remaining_gas: RemainingGas,
-        state: &mut InterpreterState<'_, T>,
-    ) -> InstrFnRet
-);
+pub(super) type RawInstrFn<T> = imp::RawInstrFn<T>;
 
 /// Normal instruction dispatch table.
 pub(super) type RawInstrTable<T> = [RawInstrFn<T>; 256];
@@ -112,173 +110,4 @@ where
     }
 
     crate::for_each_spec!([] make_tables)
-}
-
-extern_table! {
-    fn dispatch<
-        T: EvmTypes,
-        C: EvmConfig<T>,
-        M: InspectMode<T>,
-        const OP: u8,
-        const DYNAMIC_GAS: bool,
-    >(
-        pc: Pc,
-        stack: Stack<'_>,
-        remaining_gas: RemainingGas,
-        state: &mut InterpreterState<'_, T>,
-    ) -> InstrFnRet {
-        dispatch_mono::<T, C, M, DYNAMIC_GAS>(pc, stack, remaining_gas, state, OP)
-    }
-}
-
-#[cold] // Not cold, but avoids MIR inlining.
-#[inline(always)]
-fn dispatch_mono<T: EvmTypes, C: EvmConfig<T>, M: InspectMode<T>, const DYNAMIC_GAS: bool>(
-    mut pc: Pc,
-    mut stack: Stack<'_>,
-    mut remaining_gas: RemainingGas,
-    state: &mut InterpreterState<'_, T>,
-    op: u8,
-) -> InstrFnRet {
-    let initial_remaining_gas = remaining_gas;
-    let instr = C::VERSION_TABLES.instruction(op).instr;
-    if M::INSPECT {
-        M::step(state, pc, stack.len);
-    }
-    let r;
-    match pre_step::<T, C>(&mut remaining_gas, op) {
-        Ok(()) => {
-            if M::INSPECT || DYNAMIC_GAS {
-                state.gas_mut().set_remaining(remaining_gas.get());
-            }
-            r = instr(&mut pc, stack.as_mut(), state);
-            if DYNAMIC_GAS {
-                remaining_gas.set(state.gas_mut().remaining());
-            }
-            if !M::INSPECT || r.is_ok() {
-                super::inc_pc(&mut pc, op);
-            }
-        }
-        Err(e) => {
-            if M::INSPECT {
-                state.gas_mut().set_remaining(remaining_gas.get());
-            }
-            r = Err(e);
-        }
-    }
-    if M::INSPECT {
-        state.set_result(r);
-        M::step_end(state, pc, stack.len);
-    }
-    if r.is_err() {
-        cold_path();
-        if !M::INSPECT {
-            state.set_result(r);
-        }
-        let stack_len = if stack.len <= STACK_LIMIT { stack.len } else { 0 };
-        return dispatch_return(
-            Pc::new(core::ptr::null()),
-            initial_remaining_gas.get().wrapping_sub(remaining_gas.get()),
-            stack_len,
-        );
-    }
-    dispatch_return(pc, initial_remaining_gas.get().wrapping_sub(remaining_gas.get()), stack.len)
-}
-
-#[inline(always)]
-const fn pre_step<T: EvmTypes, C: EvmConfig<T>>(
-    remaining_gas: &mut RemainingGas,
-    op: u8,
-) -> Result {
-    remaining_gas.spend(C::VERSION_TABLES.static_gas(op) as _)
-}
-
-const _: () = assert!(STACK_LIMIT <= (1 << STACK_LEN_BITS));
-
-#[cfg(dispatch_packed)]
-type GasStackLen = PackedGasStackLen;
-
-#[cfg(not(dispatch_packed))]
-type GasStackLen = UnpackedGasStackLen;
-
-const STACK_LEN_BITS: u32 = 11;
-
-#[cfg(dispatch_packed)]
-const GAS_BITS: u32 = usize::BITS - STACK_LEN_BITS;
-#[cfg(dispatch_packed)]
-const GAS_MASK: usize = usize::MAX >> STACK_LEN_BITS;
-
-#[cfg(dispatch_single_return)]
-#[derive(Clone, Copy)]
-#[repr(C)]
-pub(crate) struct DispatchResult {
-    pc: Pc,
-    gas_stack_len: GasStackLen,
-}
-
-#[inline(always)]
-const fn dispatch_return(pc: Pc, gas_spent: u64, stack_len: usize) -> InstrFnRet {
-    let gas_stack_len = GasStackLen::new(gas_spent, stack_len);
-    #[cfg(not(dispatch_single_return))]
-    {
-        (pc, gas_stack_len)
-    }
-    #[cfg(dispatch_single_return)]
-    {
-        DispatchResult { pc, gas_stack_len }
-    }
-}
-
-#[inline(always)]
-pub(crate) const fn unpack_ret(ret: InstrFnRet) -> (Pc, u64, usize) {
-    #[cfg(not(dispatch_single_return))]
-    let (pc, gas_stack_len) = ret;
-    #[cfg(dispatch_single_return)]
-    let DispatchResult { pc, gas_stack_len } = ret;
-
-    let (gas_spent, stack_len) = gas_stack_len.unpack();
-    (pc, gas_spent, stack_len)
-}
-
-/// Normal dispatch gas spent and stack length, packed into one word on 64-bit native targets.
-#[cfg(dispatch_packed)]
-#[derive(Clone, Copy)]
-#[repr(transparent)]
-pub(crate) struct PackedGasStackLen(usize);
-
-#[cfg(dispatch_packed)]
-impl PackedGasStackLen {
-    #[inline(always)]
-    pub(crate) const fn new(gas_spent: u64, stack_len: usize) -> Self {
-        debug_assert!(stack_len <= STACK_LIMIT && gas_spent as usize <= GAS_MASK);
-        Self((stack_len << GAS_BITS) | (gas_spent as usize & GAS_MASK))
-    }
-
-    #[inline(always)]
-    pub(crate) const fn unpack(self) -> (u64, usize) {
-        ((self.0 & GAS_MASK) as u64, self.0 >> GAS_BITS)
-    }
-}
-
-/// Normal dispatch gas spent and stack length for targets where packing into `usize` is unsuitable.
-#[cfg(not(dispatch_packed))]
-#[derive(Clone, Copy)]
-#[repr(C)]
-pub(crate) struct UnpackedGasStackLen {
-    gas_spent: u64,
-    stack_len: usize,
-}
-
-#[cfg(not(dispatch_packed))]
-impl UnpackedGasStackLen {
-    #[inline(always)]
-    pub(crate) const fn new(gas_spent: u64, stack_len: usize) -> Self {
-        debug_assert!(stack_len <= STACK_LIMIT);
-        Self { gas_spent, stack_len }
-    }
-
-    #[inline(always)]
-    pub(crate) const fn unpack(self) -> (u64, usize) {
-        (self.gas_spent, self.stack_len)
-    }
 }
