+++ scripts/dump_asm.py
@@ -34,6 +34,9 @@ OPCODE_RS = ROOT / "crates" / "evm2" / "src" / "interpreter" / "opcode.rs"
 DEFAULT_OUT = ROOT / "tmp" / "dump"
 DISPATCH_SYMBOLS = (
     "evm2::interpreter::instructions::table::normal::dispatch::<",
+    "evm2::interpreter::instructions::table::normal::packed::dispatch::<",
+    "evm2::interpreter::instructions::table::normal::single_return::dispatch::<",
+    "evm2::interpreter::instructions::table::normal::unpacked::dispatch::<",
     "evm2::interpreter::instructions::table::tco::tail_dispatch::<",
 )
 DISPATCH_OPCODE = re.compile(r",\s*(\d+)(?:,\s*(?:true|false))*?>")
